convert fully to the new system
