+++ src/bin/yagbas.rs
@@ -1,8 +1,5 @@
-use yagbas::tokenizer::tokenize;
-
-const SOURCE: &str = include_str!("../../tests/all_tests/minimum_program.yag");
+const _SOURCE: &str = include_str!("../../tests/all_tests/minimum_program.yag");
 
 fn main() {
-  let tokens = tokenize(SOURCE).collect::<Vec<_>>();
-  println!("{tokens:?}");
+  println!("hello")
 }
