+++ src/lib.rs
@@ -12,4 +12,3 @@
 
 pub mod parser;
 pub mod tokenizer;
-pub mod tokenizer2;
